# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Challenge API Reference",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/trucks/{id}": {
            "put": {
                "tags": [
                    "TRUCKS"
                ],
                "summary": "Update the data for a particular truck. The identifier (ID) sent as a parameter corresponds to the truck in question.",
                "operationId": "App\\Http\\Documentation\\TruckDocumentation::update",
                "parameters": [
                    {
                        "name": "name",
                        "in": "query",
                        "description": "The name of the truck.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "city",
                        "in": "query",
                        "description": "The city.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "state",
                        "in": "query",
                        "description": "The state.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "zip",
                        "in": "query",
                        "description": "The zip code.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Process successfully completed"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "TRUCKS"
                ],
                "summary": "Delete the data for a particular truck. The identifier (ID) sent as a parameter corresponds to the truck in question.",
                "operationId": "App\\Http\\Documentation\\TruckDocumentation::destroy",
                "responses": {
                    "200": {
                        "description": "process successfully completed"
                    },
                    "default": {
                        "description": "Unexpected error"
                    }
                }
            }
        }
    }
}